+++ src/helper.ts
@@ -4,30 +4,26 @@ export interface QiankunProps {
 }
 
 export interface QiankunLifeCycle {
-  bootstrap: (props?: QiankunProps) => void | Promise<void>;
-  mount: (props: QiankunProps) => void | Promise<void>;
-  unmount: (props: QiankunProps) => void | Promise<void>;
-  update: (props: QiankunProps) => void | Promise<void>;
+  bootstrap?: (props?: QiankunProps) => void | Promise<void>;
+  mount?: (props: QiankunProps) => void | Promise<void>;
+  unmount?: (props: QiankunProps) => void | Promise<void>;
+  update?: (props: QiankunProps) => void | Promise<void>;
 }
 
 export interface QiankunWindow {
   __POWERED_BY_QIANKUN__?: boolean;
+  __INJECTED_PUBLIC_PATH_BY_QIANKUN__?: string;
+  qiankunLifeCycles?: QiankunLifeCycle;
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   [x: string]: any;
 }
 
-export const qiankunWindow: QiankunWindow = typeof window !== 'undefined' ? window.proxy || window : {};
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export const qiankunWindow: QiankunWindow = typeof window !== 'undefined' ? (window as any).proxy || window : {};
 
-export const renderWithQiankun = (qiankunLifeCycle: QiankunLifeCycle) => {
-  // 函数只有一次执行机会，需要把生命周期赋值给全局
+export const exportQiankunLifeCycles = (qiankunLifeCycle: QiankunLifeCycle) => {
+  // The function has only one chance to execute, and the lifecycle needs to be assigned to the global scope.
   if (qiankunWindow?.__POWERED_BY_QIANKUN__) {
-    if (!window.moudleQiankunAppLifeCycles) {
-      window.moudleQiankunAppLifeCycles = {};
-    }
-    if (qiankunWindow.qiankunName) {
-      window.moudleQiankunAppLifeCycles[qiankunWindow.qiankunName] = qiankunLifeCycle;
-    }
+    qiankunWindow.qiankunLifeCycles = qiankunLifeCycle;
   }
 };
-
-export default renderWithQiankun;